+++ Repast_ParaRoute_PartOpt/EvacSim/src/evacSim/NetworkEventHandler.java
@@ -180,7 +180,7 @@ public class NetworkEventHandler {
 						case 1: // Change speed limit
 							if (!road.checkEventFlag()) {
 								road.setDefaultFreeSpeed();
-								road.updateFreeFlowSpeed_event(event.value1);
+								road.updateFreeFlowSpeed_event((float) event.value1);
 								road.setEventFlag();
 								return event;
 							} else {
@@ -219,7 +219,7 @@ public class NetworkEventHandler {
 					// Restore the event
 					switch (event.eventID) {
 						case 1: // restore speed limit
-							road.updateFreeFlowSpeed_event(road.getDefaultFreeSpeed()); // To be moved into a buffer variable in the road
+							road.updateFreeFlowSpeed_event((float) road.getDefaultFreeSpeed()); // To be moved into a buffer variable in the road
 							road.restoreEventFlag();
 							return event;
 						// Other cases to be implemented later
